chore: redid compose item models
