+++ nextjs/tsconfig.json
@@ -27,7 +27,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "../autumn.config.ts"
+    "../cli/autumn.config.ts"
   ],
   "exclude": ["node_modules"]
 }
