+++ package/src/compose/index.ts
@@ -5,25 +5,11 @@ import {
   feature,
   featureItem,
   pricedFeatureItem,
-} from "./builders/core";
-import {
-  ProductItemInterval,
-  Feature,
-  Product,
-  ProductItem,
-  UsageModel,
-} from "./models/core";
+} from "./builders/builderFunctions";
+import { Feature, Product } from "./models/composeModels";
+import { ProductItem } from "./models/productItemModels";
 
-export {
-  product,
-  auth,
-  priceItem,
-  ProductItemInterval,
-  feature,
-  featureItem,
-  pricedFeatureItem,
-  UsageModel,
-};
+export { product, auth, priceItem, feature, featureItem, pricedFeatureItem };
 
 export type { Feature, Product, ProductItem };
 export type Infinity = "infinity";
